Update 'create_building_cadastre_wizard' wizard
